Move functions to the new common.ts
