+++ lib/modules/manager/pip-compile/common.ts
@@ -3,9 +3,10 @@ import { split } from 'shlex';
 import { logger } from '../../../logger';
 import type { ExecOptions } from '../../../util/exec/types';
 import { ensureCacheDir } from '../../../util/fs';
+import * as hostRules from '../../../util/host-rules';
 import { regEx } from '../../../util/regex';
-import type { UpdateArtifactsConfig } from '../types';
-import type { PipCompileArgs } from './types';
+import type { PackageFileContent, UpdateArtifactsConfig } from '../types';
+import type { GetRegistryUrlVarsResult, PipCompileArgs } from './types';
 
 export function getPythonConstraint(
   config: UpdateArtifactsConfig,
@@ -218,4 +219,58 @@ function throwForUnknownOption(arg: string): void {
     return;
   }
   throw new Error(`Option ${arg} not supported (yet)`);
+}
+
+function buildRegistryUrl(url: string): URL {
+  const hostRule = hostRules.find({ url });
+  const ret = new URL(url);
+  if (!ret.username) {
+    ret.username = hostRule.username ?? '';
+    ret.password = hostRule.password ?? '';
+  }
+  return ret;
+}
+
+function getRegistryUrlVarFromUrls(
+  varName: keyof GetRegistryUrlVarsResult['environmentVars'],
+  urls: URL[],
+): GetRegistryUrlVarsResult {
+  let haveCredentials = false;
+  for (const url of urls) {
+    if (url.username) {
+      haveCredentials = true;
+    }
+  }
+  const registryUrlsString = urls.map((url) => url.href).join(' ');
+  const ret: GetRegistryUrlVarsResult = {
+    haveCredentials,
+    environmentVars: {},
+  };
+  if (registryUrlsString) {
+    ret.environmentVars[varName] = registryUrlsString;
+  }
+  return ret;
+}
+
+export function getRegistryUrlVarsFromPackageFile(
+  packageFile: PackageFileContent | null,
+): GetRegistryUrlVarsResult {
+  // There should only ever be one element in registryUrls, since pip_requirements gets them from --index-url
+  // flags in the input file, and that only makes sense once
+  const indexUrl = getRegistryUrlVarFromUrls(
+    'PIP_INDEX_URL',
+    packageFile?.registryUrls?.map(buildRegistryUrl) ?? [],
+  );
+  const extraIndexUrls = getRegistryUrlVarFromUrls(
+    'PIP_EXTRA_INDEX_URL',
+    packageFile?.additionalRegistryUrls?.map(buildRegistryUrl) ?? [],
+  );
+
+  return {
+    haveCredentials: indexUrl.haveCredentials || extraIndexUrls.haveCredentials,
+    environmentVars: {
+      ...indexUrl.environmentVars,
+      ...extraIndexUrls.environmentVars,
+    },
+  };
 }
